remove to_position; don't need it
It's not useful to fill in gl_Position.w with y[3]. Filling w with y[3]
was the old behavior when the user did not do stereo at the end of a 4+
dimensional pipe. However, this was not a concious decision, and doesn't
make much intuitive sense, since w only affects scaling.