update default audio.hue
Whole concept probably needs a revisit when I finally get around to
fixing audio. But for now setting it to 1 is enough to rationalize it.
The way it's used in practice (hue * abs(p[0])) means that it basically
scales the range of colors that appear in the scene. However, when I
spend more time with audio in the future, I'll probably think of more
interesting ways to use it.